backend_balap_in/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.2
django-cors-headers==4.6.0
django-cron==0.6.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
geographiclib==2.0
geopy==2.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
msgpack==1.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.5
numpy==2.1.3
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.1
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.1
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.2
